Move sign types to object builder api
