+++ fabric-object-builder-api-v1/src/client/resources/fabric-object-builder-v1.client.mixins.json
@@ -3,8 +3,10 @@
   "package": "net.fabricmc.fabric.mixin.object.builder.client",
   "compatibilityLevel": "JAVA_16",
   "client": [
+    "EntityModelLayersMixin",
     "ModelPredicateProviderRegistryAccessor",
-    "ModelPredicateProviderRegistrySpecificAccessor"
+    "ModelPredicateProviderRegistrySpecificAccessor",
+    "TexturedRenderLayersMixin"
   ],
   "injectors": {
     "defaultRequire": 1
