+++ fabric-object-builder-api-v1/src/main/java/net/fabricmc/fabric/api/object/builder/v1/sign/SignTypeRegistry.java
@@ -28,7 +28,10 @@ import net.fabricmc.fabric.mixin.object.builder.SignTypeAccessor;
  *
  * <P>These textures are stored under {@code namespace/textures/entity/signs/}.
  */
-public class SignTypeRegistry {
+public final class SignTypeRegistry {
+	private SignTypeRegistry() {
+	}
+
 	/**
 	 * Creates and registers a {@link SignType}.
 	 *
