+++ fabric-object-builder-api-v1/src/testmod/java/net/fabricmc/fabric/test/object/builder/TealSignTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.test.rendering;
+package net.fabricmc.fabric.test.object.builder;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
