+++ fabric-object-builder-api-v1/src/testmod/resources/fabric-object-builder-api-v1-testmod.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.test.rendering.mixin",
+  "package": "net.fabricmc.fabric.test.object.builder.mixin",
   "compatibilityLevel": "JAVA_17",
   "client": [
     "BlockEntityRendererFactoriesAccessor"
