+++ fabric-object-builder-api-v1/src/testmod/resources/fabric.mod.json
@@ -20,12 +20,19 @@
     "fabric-object-builder-api-v1": "*"
   },
   "description": "Test mod for fabric object builder API v1.",
+  "mixins": [
+    "fabric-object-builder-api-v1-testmod.mixins.json"
+  ],
   "entrypoints": {
     "main": [
       "net.fabricmc.fabric.test.object.builder.BlockEntityTypeBuilderTest",
       "net.fabricmc.fabric.test.object.builder.CriterionRegistryTest::init",
       "net.fabricmc.fabric.test.object.builder.VillagerTypeTest1",
-      "net.fabricmc.fabric.test.object.builder.VillagerTypeTest2"
+      "net.fabricmc.fabric.test.object.builder.VillagerTypeTest2",
+      "net.fabricmc.fabric.test.object.builder.TealSignTest"
+    ],
+    "client": [
+      "net.fabricmc.fabric.test.object.builder.client.TealSignClientTest"
     ]
   }
 }
