--- fabric-rendering-v1/src/client/java/net/fabricmc/fabric/mixin/client/rendering/TexturedRenderLayersMixin.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.mixin.client.rendering;
-
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
-
-import net.minecraft.client.render.TexturedRenderLayers;
-
-@Mixin(TexturedRenderLayers.class)
-public class TexturedRenderLayersMixin {
-	@ModifyArg(method = "createSignTextureId", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/Identifier;<init>(Ljava/lang/String;)V"))
-	private static String modifyTextureId(String id) {
-		if (id.contains(":")) {
-			String[] split = id.replace("entity/signs/", "").split(":");
-			split[1] = ":entity/signs/" + split[1];
-			return split[0] + split[1];
-		}
-
-		return id;
-	}
-}
