+++ fabric-rendering-v1/src/testmod/resources/fabric.mod.json
@@ -8,19 +8,14 @@
   "depends": {
     "fabric-rendering-v1": "*"
   },
-  "mixins": [
-    "fabric-rendering-v1-testmod.mixins.json"
-  ],
   "entrypoints": {
     "main": [
-      "net.fabricmc.fabric.test.rendering.TooltipComponentTestInit",
-      "net.fabricmc.fabric.test.rendering.TealSignTest"
+      "net.fabricmc.fabric.test.rendering.TooltipComponentTestInit"
     ],
     "client": [
       "net.fabricmc.fabric.test.rendering.client.WorldRenderEventsTests",
       "net.fabricmc.fabric.test.rendering.client.ArmorRenderingTests",
       "net.fabricmc.fabric.test.rendering.client.FeatureRendererTest",
-      "net.fabricmc.fabric.test.rendering.client.TealSignClientTest",
       "net.fabricmc.fabric.test.rendering.client.TooltipComponentTests",
       "net.fabricmc.fabric.test.rendering.client.DimensionalRenderingTest"
     ]
